Hibernate recorder process after compact cycle
The `Recorder` process "touches" large batches of JSON received from
`Reporter` processes, but it doesn't operate on the data often enough to
trigger a full GC. The entire mechanics are explained in [this
post][post] on the ElixirForum.

Now, the `Recorder` hibernates after compacting to trigger a fullsweep
garbage collection.

Closes #1

[post]: https://elixirforum.com/t/extremely-high-memory-usage-in-genservers/4035/23