Prevent excessive message queue in OTP 26.2.5
A bug fix in OTP 26.2.5 made the performance of `:ets.select/3` with a
map in the key much slower, effectively linear with the number of
objects in the able. That performance degredation is enough to
bottleneck the `Oban.Met.Recorder` message queue, which bloats the
process and may cause OOM errors for active systems.

This restores, and even improves, the performance of `:ets.select/3` for
recorder operations by moving label maps out of the recorded oject key.

Closes sorentwo/oban#1079